// payroll-service-api/src/grpc/generated/employee.ts
export interface EmployeeResponse {
  employeeId: number;
  firstName: string;
  lastName: string;
  email: string;
  salary: number;
  jobId: number;
  departmentId: number;
  employmentStatus: string;
  employmentType: string;
  hireDate: string;
  updatedAt: string;
}